Plugin - Mss Controller - Moved lib components to new lib projects
