+++ software/plugins/mss-controller/mss-controller-plugin-devtools/deployment/pom.xml
@@ -11,6 +11,11 @@
   </parent>
   <dependencies>
     <dependency>
+      <groupId>tech.ebp.oqm.plugin</groupId>
+      <artifactId>mss-controller-plugin-lib</artifactId>
+      <version>${project.parent.version}</version>
+    </dependency>
+    <dependency>
       <groupId>io.quarkus</groupId>
       <artifactId>quarkus-arc-deployment</artifactId>
     </dependency>
