+++ software/plugins/mss-controller/mss-controller-plugin-devtools/deployment/src/main/java/tech/ebp/oqm/plugin/mssController/devTools/deployment/testModules/MssTestModule.java
@@ -1,7 +1,7 @@
 package tech.ebp.oqm.plugin.mssController.devTools.deployment.testModules;
 
-import tech.ebp.oqm.plugin.mssController.devTools.runtime.model.command.MssCommand;
-import tech.ebp.oqm.plugin.mssController.devTools.runtime.model.command.response.ModuleInfo;
+import tech.ebp.oqm.plugin.mssController.lib.command.MssCommand;
+import tech.ebp.oqm.plugin.mssController.lib.command.response.ModuleInfo;
 
 public class MssTestModule {
