+++ software/plugins/mss-controller/mss-controller-plugin-devtools/deployment/src/main/java/tech/ebp/oqm/plugin/mssController/devTools/deployment/testModules/MssTestSerialModule.java
@@ -1,6 +1,6 @@
 package tech.ebp.oqm.plugin.mssController.devTools.deployment.testModules;
 
-import tech.ebp.oqm.plugin.mssController.devTools.runtime.model.command.response.ModuleInfo;
+import tech.ebp.oqm.plugin.mssController.lib.command.response.ModuleInfo;
 
 public class MssTestSerialModule extends MssTestModule {
 	protected MssTestSerialModule(ModuleInfo moduleInfo) {
