--- software/plugins/mss-controller/mss-controller-plugin-devtools/runtime/src/main/java/tech/ebp/oqm/plugin/mssController/devTools/runtime/model/command/response/CommandResponseStatus.java
@@ -1,7 +1,0 @@
-package tech.ebp.oqm.plugin.mssController.devTools.runtime.model.command.response;
-
-public enum CommandResponseStatus {
-	OK,
-	R_ERR,
-	ERR
-}
