+++ software/plugins/mss-controller/mss-controller-plugin-lib/lombok.config
@@ -1,0 +1,4 @@
+lombok.accessors.chain=true
+lombok.addLombokGeneratedAnnotation = true
+#lombok.addJavaxGeneratedAnnotation = true
+lombok.builder.className = Builder
