+++ software/plugins/mss-controller/mss-controller-plugin-lib/src/main/java/tech/ebp/oqm/plugin/mssController/lib/command/IdentifyModCommand.java
@@ -1,4 +1,4 @@
-package tech.ebp.oqm.plugin.mssController.devTools.runtime.model.command;
+package tech.ebp.oqm.plugin.mssController.lib.command;
 
 import lombok.AccessLevel;
 import lombok.Data;
