+++ software/plugins/mss-controller/mss-controller-plugin-lib/src/main/java/tech/ebp/oqm/plugin/mssController/lib/command/MssCommand.java
@@ -1,4 +1,4 @@
-package tech.ebp.oqm.plugin.mssController.devTools.runtime.model.command;
+package tech.ebp.oqm.plugin.mssController.lib.command;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
