+++ software/plugins/mss-controller/mss-controller-plugin-lib/src/main/java/tech/ebp/oqm/plugin/mssController/lib/command/response/CommandResponse.java
@@ -1,4 +1,4 @@
-package tech.ebp.oqm.plugin.mssController.devTools.runtime.model.command.response;
+package tech.ebp.oqm.plugin.mssController.lib.command.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
