+++ software/plugins/mss-controller/mss-controller-plugin-lib/src/main/java/tech/ebp/oqm/plugin/mssController/lib/command/response/CommandResponseStatus.java
@@ -1,0 +1,7 @@
+package tech.ebp.oqm.plugin.mssController.lib.command.response;
+
+public enum CommandResponseStatus {
+	OK,
+	R_ERR,
+	ERR
+}
